refactor: move SearchFilters,ActiveFilters,SearchFilterCheckbox
