+++ src/components/TeacherComponents/SearchActiveFilters/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./SearchActiveFilters";
