+++ src/components/TeacherComponents/SearchFilterCheckbox/SearchFilterCheckbox.test.tsx
@@ -1,6 +1,7 @@
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
-
 import SearchFilterCheckbox from "./SearchFilterCheckbox";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
 
 const props = {
   slug: "ks1",
