+++ src/components/TeacherComponents/SearchFilters/SearchFilters.test.tsx
@@ -1,7 +1,8 @@
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
 
 import SearchFilters from "./SearchFilters";
 import { searchFilters, mockOnChange } from "./test-helpers";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 
 jest.mock("next/dist/client/router", () => require("next-router-mock"));
